ס

+++ 05-Node/99-Assignments/Raziel/public/index.js
@@ -30,9 +30,6 @@ let ENTRY_LIST=[];
 let balance = 0, income = 0, outcome = 0;
 const DELETE = "delete", EDIT = "edit";
 
-// // LOOK IF THERE IS SAVED DATA IN LOCALSTORAGE
-// ENTRY_LIST = JSON.parse(localStorage.getItem("entry_list")) || [];
-// updateUI();
 
 // EVENT LISTENERS
 expenseBtn.addEventListener("click",function(){
@@ -59,7 +56,7 @@ addExpense.addEventListener("click",addExpense1)
 
 async function addExpense1(){
  try {
-    if(!expenseTitle.value || !expenseAmount.value ) return;
+    // if(!expenseTitle.value || !expenseAmount.value ) return;
     let expense = {
         type : "expense",
         title : expenseTitle.value,
@@ -82,63 +79,10 @@ console.log(ENTRY_LIST);
 }
 
 
-    // IF ONE OF THE INPUTS IS EMPTY => EXIT
-//     try {
-//         if(!expenseTitle.value || !expenseAmount.value ) return;
-
-//         // SAVE THE ENTRY TO ENTRY_LIST
-//         let expense = {
-//             type : "expense",
-//             title : expenseTitle.value,
-//             amount : parseInt(expenseAmount.value)
-//         }
-//         ENTRY_LIST.push(expense);
-        
-//        const data = {type : "expense",tile :expenseTitle.value,amount : parseInt(expenseAmount.value)}
-//        const options={
-//            method:'POST',
-//            headers: {
-//             'Content-Type': 'application/json'
-           
-//           },
-//           body: JSON.stringify(data) // body data type must match "Content-Type" header
-
-//        }
-//        fetch('/addExpense',options);
-//         updateUI();
-//         clearInput( [expenseTitle, expenseAmount] )  
-//     } catch (error) {
-//         console.error(error)
-//     }
-    
-// })
-
 addIncome.addEventListener("click",addIncome1());
-//     // IF ONE OF THE INPUTS IS EMPTY => EXIT
-//     if(!incomeTitle.value || !incomeAmount.value ) return;
 
-//     // SAVE THE ENTRY TO ENTRY_LIST
-//     let income = {
-//         type : "income",
-//         title : incomeTitle.value,
-//         amount : parseInt(incomeAmount.value)
-//     }
-//     ENTRY_LIST.push(income);
-//     const data = {type : "income",tile :incomeTitle.value,amount : parseInt(incomeAmount.value)}
-//        const options={
-//            method:'POST',
-//            headers: {
-//             'Content-Type': 'application/json'
-           
-//           },
-//           body: JSON.stringify(data) // body data type must match "Content-Type" header
-
-//        }
-//        fetch('/addIncome',options);
-//     updateUI();
-//     clearInput( [incomeTitle, incomeAmount] )
-// })
 async function addIncome1(){
+    // if(!expenseTitle.value || !expenseAmount.value ) return;
 
 try {
     let income = {
@@ -181,11 +125,19 @@ function deleteOrEdit(event){
 
 //     updateUI();
 // }
- async function  deleteEntryt(id){
+async function deleteEntery(id) {
+    try {
+        const option = confirm(`Are you sure do you want to delete this task?`);
+        if (option) {
+            const Entery= await axios.delete(`/deleteEntry/${id}`);
+          ENTRY_LIST.push(...Entery.data);
+          updateUI();
+        }
+    } catch (error) {
+        console.error(error);
+    }
+};
 
-    const res=await axios.delete(`/deleteEntry/${id}`);
-    updateUI();
- }
 function editEntry(entry){
     console.log(entry)
     let ENTRY = ENTRY_LIST[entry.id];
@@ -198,7 +150,7 @@ function editEntry(entry){
         expenseTitle.value = ENTRY.title;
     }
 
-    deleteEntry(entry);
+
 }
 
 function updateUI(){
